Change record-reference to record-inclusion
