+++ tests/jballerina-unit-test/src/test/resources/test-src/record/closed_record_type_inclusion.bal
