+++ tests/jballerina-unit-test/src/test/resources/test-src/record/open_record_type_inclusion.bal
