+++ tests/jballerina-unit-test/src/test/resources/test-src/record/open_record_type_inclusion_negative.bal
@@ -95,3 +95,17 @@ type Employee record {
     *PersonTwo;
     int age;
 };
+
+public type Foo record {
+    anydata body;
+};
+
+// Out of order inclusion test : added to test a NPE
+type Bar record {
+    *Foo;
+    Baz body;   // defined after the type definition
+};
+
+type Baz record {
+    int id;
+};
