+++ tests/jballerina-unit-test/src/test/resources/testng.xml
@@ -241,8 +241,8 @@
                     <exclude name="testAsyncNonNativeBasic11" />
                 </methods>
             </class>
-            <class name="org.ballerinalang.test.bala.record.OpenRecordTypeReferenceTest" />
-            <class name="org.ballerinalang.test.bala.record.ClosedRecordTypeReferenceTest" />
+            <class name="org.ballerinalang.test.bala.record.OpenRecordTypeInclusionTest" />
+            <class name="org.ballerinalang.test.bala.record.ClosedRecordTypeInclusionTest" />
             <class name="org.ballerinalang.test.main.function.MainFunctionsTest" />
             <class name="org.ballerinalang.test.main.function.OperandCompilationTest" />
             <class name="org.ballerinalang.test.main.function.OptionCompilationTest" />
